chrome_plugin/requirements.txt
fastapi>=0.115
uvicorn>=0.32
pyyaml>=6.0
